+++ TradeMatchingEngine/Orders/Repositories/Query/IOrderQuery.cs
@@ -4,8 +4,8 @@ namespace TradeMatchingEngine.Orders.Repositories.Query
 {
     public interface IOrderQuery
     {
-        Task<Order> GetOrderById(int id);
+        Task<Order> GetOrderById(long id);
         Task<IEnumerable<Order>> GetAllOrders();
-        Task<int> GetLastOrder();
+        Task<long> GetLastOrder();
     }
 }
